+++ src/main/components/ExampleComponent.java
@@ -1,8 +1,8 @@
 package main.components;
 
 
-import main.annotations.Autowired;
-import main.annotations.Component;
+import main.container.annotations.Autowired;
+import main.container.annotations.Component;
 
 @Component
 public
