+++ src/main/container/annotations/Component.java
@@ -1,4 +1,4 @@
-package main.annotations;
+package main.container.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,7 +6,6 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.PARAMETER)
-public @interface Qualifier {
-    String value();
+@Target(ElementType.TYPE)
+public @interface Component {
 }
