+++ src/main/container/annotations/Qualifier.java
@@ -1,4 +1,4 @@
-package main.annotations;
+package main.container.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,6 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.CONSTRUCTOR)
-public @interface Autowired {
+@Target(ElementType.PARAMETER)
+public @interface Qualifier {
+    String value();
 }
