+++ src/main/container/services/ClassScanner.java
@@ -1,4 +1,4 @@
-package main;
+package main.container.services;
 
 import java.io.File;
 import java.net.URL;
